Continuing creator tests

+++ src/main/java/mt/edu/um/b_d_app/HighRiskTransactionCreator.java
@@ -20,7 +20,7 @@ public class HighRiskTransactionCreator extends TransactionCreator {
     private int MainTransSourceAccountNo = 3143;
     
     
-    public Transaction createTransaction(Account depositDSTAccount, float depositAmount, ArrayList<Account> destinationList, ArrayList<Double> amountList,AccountDatabase database)
+    public Transaction createTransaction(Account depositDSTAccount, double depositAmount, ArrayList<Account> destinationList, ArrayList<Double> amountList,AccountDatabase database)
     {
         //Specification states that all of the compound transactions have three transactions at at level 1 (topmost level of the transaction)
         Transaction result = new CompositeTransaction("HighRiskTransaction");
