+++ src/main/java/mt/edu/um/b_d_app/TransactionCreator.java
@@ -8,7 +8,7 @@ public class TransactionCreator {
 
     //transaction manager will receive list of account numbers and find the accounts
     //move database back to transaction manager
-    public Transaction createTransaction(RiskTypes type,Account depositDSTAccount, float depositAmount ,ArrayList<Account> destinationList, ArrayList<Double> amountList, AccountDatabase database){
+    public Transaction createTransaction(RiskTypes type,Account depositDSTAccount, double depositAmount ,ArrayList<Account> destinationList, ArrayList<Double> amountList, AccountDatabase database){
         
         TransactionCreator creator = findCreatorForTypeOfTransaction(type);
         
@@ -27,7 +27,7 @@ public class TransactionCreator {
         return null;//set to null as no default type was specified in the assignment specification
     }
     
-    public Transaction createTransaction(Account depositDSTAccount, float depositAmount, ArrayList<Account> destinationList, ArrayList<Double> amountList,AccountDatabase database){
+    public Transaction createTransaction(Account depositDSTAccount, double depositAmount, ArrayList<Account> destinationList, ArrayList<Double> amountList,AccountDatabase database){
         
         return null;//no type specified for this level in the transaction creator hierarchy
